backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]
reportlab==3.6.12
boto3==1.34.146
celery[redis]==5.3.6
redis==5.0.1
jinja2>=3.1.0
